Added new resource system.

+++ Editor/src/EditorWindow/InspectorWindow/EntityInspector/EntityInspector.cpp
@@ -93,7 +93,7 @@ void EntityInspector::Procedure(const VL::Event& e)
 		{
 			auto pointLight = scene->GetComponent<VL::PointLight>(selectedEntity, i);
 
-			std::string label = "##Color" + std::to_string((int)pointLight.get());
+			std::string label = "##Color" + std::to_string((uint64_t)pointLight.get());
 			ImGui::ColorPicker3(label.c_str(), glm::value_ptr(pointLight->Color), ImGuiColorEditFlags_Float);
 
 			Utils::ImGuiFloat("Brightness", &pointLight->Brightness);
